insert and update user

+++ marring-dao/src/main/java/cn/marring/dao/enums/Gender.java
@@ -1,22 +1,23 @@
 package cn.marring.dao.enums;
 
+
 /**
  * @author Wn 2020-06-30 14:27
  */
-public enum  Gender {
+public enum Gender {
     UNKNOW(0, "未知"),
     MAN(1, "先生"),
     WOMAN(2, "女士");
 
-    private Byte value;
+    private int value;
     private String name;
 
     Gender(int value, String name) {
-        this.value = (byte)value;
+        this.value = (byte) value;
         this.name = name;
     }
 
-    public Byte getValue() {
+    public int getValue() {
         return this.value;
     }
